+++ packages/winnow/src/query.js
@@ -3,17 +3,20 @@ const Geometry = require('./geometry')
 const Where = require('./where')
 const Select = require('./select')
 const Order = require('./order')
+const GroupBy = require('./groupBy')
 
 function create (options) {
   let query = Select.createClause(options)
   const where = Where.createClause(options)
   const geomFilter = Geometry.createClause(options)
   const order = Order.createClause(options)
+  const groupBy = GroupBy.createClause(options)
   if (options.where || options.geometry) query += ' WHERE '
   if (options.where) query += where
   if (options.geometry && !where) query += geomFilter
   if (options.geometry && where) query += ` AND ${geomFilter}`
   // if (options.aggregates) return query
+  if (options.groupBy && groupBy) query += groupBy
   if (options.order || options.orderByFields) query += order
   if (options.limit) query += ` LIMIT ${options.limit}`
   if (options.offset) query += ` OFFSET ${options.offset}` // handled in executeQuery.js